Support for --update-existing option
To ensure that I don't accidentally add missing columns to an existing table, I added a CLI option for `--update-existing` that defaults to false.

I also documented in the README how to pass options like that when testing locally.

Lastly, I commented out a console log for "API Response" that was generating way too much noise.